Move error handling out of cli
